Split out atomic push functions into their own source file.
Create a Push class as a subclass of Atomic and move all of the
functions and CLI to this file.
