+++ atomic
@@ -42,6 +42,7 @@ from Atomic import install
 from Atomic import help as Help
 from Atomic import run
 from Atomic import ps
+from Atomic import push
 from Atomic import uninstall
 from Atomic.util import write_err, NoDockerDaemon, add_opt
 import traceback
@@ -136,12 +137,7 @@ def create_parser(help_text):
     images.cli(subparser)
     mount.cli(subparser)
     ps.cli(subparser)
-
-    # atomic push
-    pushp = subparser.add_parser(
-        "push", aliases=['upload'], help=_("push latest image to repository"),
-        epilog="push the latest specified image to a repository.")
-    pushp.set_defaults(func='push')
+    push.cli(subparser)
 
     # atomic pull
     pullp = subparser.add_parser("pull", help=_("pull latest image from a repository"),
@@ -149,60 +145,6 @@ def create_parser(help_text):
     pullp.add_argument("--storage", dest="backend", help=_("Specify the storage."))
     pullp.add_argument("image", help=_("image id"))
     pullp.set_defaults(func='pull_image')
-
-    # making it so we cannot call both the --pulp and --satellite commands
-    # at the same time (mutually exclusive)
-    pushgroup = pushp.add_mutually_exclusive_group()
-    pushgroup.add_argument("--pulp",
-                           default=False,
-                           action="store_true",
-                           help=_("push image using pulp"))
-    pushgroup.add_argument("--satellite",
-                           default=False,
-                           action="store_true",
-                           help=_("push image using Satellite"))
-
-    pushp.add_argument("--verify_ssl",
-                         default=None,
-                         action="store_true",
-                         help=_("flag to verify ssl of registry"))
-    pushp.add_argument("--debug",
-                         default=None,
-                         action="store_true",
-                         help=_("debug mode"))
-    pushp.add_argument("-U", "--url",
-                         dest="url",
-                         default=None,
-                         help=_("URL for remote registry"))
-    pushp.add_argument("-u", "--username",
-                         default=None,
-                         dest="username",
-                         help=_("Username for remote registry"))
-    pushp.add_argument("-p", "--password",
-                         default=None,
-                         dest="password",
-                         help=_("Password for remote registry"))
-    pushp.add_argument("image", help=_("container image"))
-    pushp.add_argument("-a", "--activation_key",
-                         default=None,
-                         dest="activation_key",
-                         help=_("Activation Key"))
-    pushp.add_argument("-r", "--repository_id",
-                         default=None,
-                         dest="repo_id",
-                         help=_("Repository ID"))
-    # pushp.add_argument("--activation_key_name",
-    #                      default=None,
-    #                      dest="activation_key_name",
-    #                      help=_("Activation Key Name"))
-    # pushp.add_argument("--repo_name", "--repository_name",
-    #                      default=None,
-    #                      dest="repo_name",
-    #                      help=_("Repository Name"))
-    # pushp.add_argument("--org_name", "--organization_name",
-    #                      default=None,
-    #                      dest="org_name",
-    #                      help=_("Organization Name"))
 
     run.cli(subparser, atomic)
     scan.cli(subparser)
